[Lib] Separate Lib subcommands
